Updated common directory and changed all refereces org.rawkintrevo._ to org.apache.mahout._
