+++ examples/cylons/bin/examples/drone-example.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 java -cp $CYLON_HOME/examples/target/examples-1.0-SNAPSHOT-jar-with-dependencies.jar \
- org.rawkintrevo.cylon.examples.localfacialrecognition.SimpleFacialRecognitionApp \
+ org.apache.mahout.cylon-example.examples.localfacialrecognition.SimpleFacialRecognitionApp \
  -c $OPEN_CV/data/haarcascades/haarcascade_frontalface_default.xml \
 	-e $CYLON_HOME/data/eigenfaces-130_2.11/eigenfaces.mmat \
 	-m $CYLON_HOME/data/eigenfaces-130_2.11/colMeans.mmat \
