+++ examples/cylons/bin/examples/producers/f2v-on-drone.sh
@@ -2,7 +2,7 @@
 
 
 java -cp $CYLON_HOME/examples/target/examples-1.0-SNAPSHOT-jar-with-dependencies.jar \
- org.rawkintrevo.cylon.examples.FacesToVectorsApp \
+ org.apache.mahout.cylon-example.examples.FacesToVectorsApp \
  -c $OPEN_CV/data/haarcascades/haarcascade_frontalface_default.xml \
 	-e $CYLON_HOME/data/eigenfaces-130_2.11/eigenfaces.mmat \
 	-m $CYLON_HOME/data/eigenfaces-130_2.11/colMeans.mmat \
