+++ examples/cylons/bin/examples/producers/simple-on-drone.sh
@@ -5,7 +5,7 @@ you built OpenCV 3.x and Solr is running on localhost:8983 with the collection '
 AND MAKE SURE YOUR DRONE IS ON AND CONNECTED!"
 
 java -cp $CYLON_HOME/examples/target/examples-1.0-SNAPSHOT-jar-with-dependencies.jar \
- org.rawkintrevo.cylon.examples.localfacialrecognition.SimpleFacialRecognitionApp \
+ org.apache.mahout.cylon-example.examples.localfacialrecognition.SimpleFacialRecognitionApp \
  -c $OPEN_CV/data/haarcascades/haarcascade_frontalface_default.xml \
 	-e $CYLON_HOME/data/eigenfaces-130_2.11/eigenfaces.mmat \
 	-m $CYLON_HOME/data/eigenfaces-130_2.11/colMeans.mmat \
