+++ examples/cylons/common/pom.xml
@@ -1,9 +1,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <parent>
-    <artifactId>cylon</artifactId>
-    <groupId>org.rawkintrevo.cylon</groupId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>cylon-example</artifactId>
+    <groupId>org.apache.mahout</groupId>
+    <version>0.13.1-SNAPSHOT</version>
+    <relativePath>../</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <artifactId>common</artifactId>
