--- examples/cylons/common/src/main/java/org/rawkintrevo/cylon/common/opencv/LoadNative.java
@@ -1,12 +1,0 @@
-package org.rawkintrevo.cylon.common.opencv;
-
-import org.opencv.core.Core;
-
-public class LoadNative {
-
-    static {
-        System.loadLibrary(Core.NATIVE_LIBRARY_NAME);
-    }
-
-    native void loadNative();
-}
