+++ examples/cylons/common/src/main/scala/org/apache/mahout/cylonexample/common/mahout/MahoutUtils.scala
@@ -1,16 +1,29 @@
-package org.rawkintrevo.cylon.common.mahout
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.mahout.cylonexample.common.mahout
 
 import java.io.{FileInputStream, FileOutputStream, ObjectInputStream, ObjectOutputStream}
 import java.nio.ByteBuffer
 
 import org.apache.mahout.math._
-import org.apache.mahout.math.algorithms.preprocessing.MeanCenter
-import org.apache.mahout.math.decompositions._
-import org.apache.mahout.math.drm._
 import org.apache.mahout.math.scalabindings._
 import org.apache.mahout.math.Matrix
 import org.apache.mahout.math.scalabindings.RLikeOps._
-import org.apache.mahout.math.drm.DrmLike
 import org.apache.mahout.math.scalabindings.MahoutCollections._
 
 import scala.collection.JavaConversions._
