+++ examples/cylons/common/src/main/scala/org/apache/mahout/cylonexample/common/solr/CylonSolrClient.scala
@@ -1,4 +1,21 @@
-package org.rawkintrevo.cylon.common.solr
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.mahout.cylonexample.common.solr
 
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
@@ -23,7 +40,7 @@ class CylonSolrClient extends Serializable{
     query.setRequestHandler("/select")
     val currentPointStr = v.toArray.mkString(",")
     val eigenfaceFieldNames = (0 until v.size()).map(i => s"e${i}_d").mkString(",")
-    val distFnStr = s"dist(2, ${eigenfaceFieldNames},${currentPointStr})"
+    val distFnStr = s"dist(2, $eigenfaceFieldNames,$currentPointStr)"
     query.setQuery("*:*")
     query.setSort(new SortClause(distFnStr, SolrQuery.ORDER.asc))
     query.setFields("name_s", "calc_dist:" + distFnStr, "last_seen_pdt")
