+++ examples/cylons/eigenfaces/pom.xml
@@ -1,9 +1,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <parent>
-    <groupId>org.rawkintrevo.cylon</groupId>
-    <artifactId>cylon</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>cylon-example</artifactId>
+    <groupId>org.apache.mahout</groupId>
+    <version>0.13.1-SNAPSHOT</version>
+    <relativePath>../</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <artifactId>spark-eigenfaces</artifactId>
@@ -77,13 +78,13 @@
 
     <!-- Our Stuff -->
     <dependency>
-      <groupId>org.rawkintrevo.cylon</groupId>
+      <groupId>org.apache.mahout.cylon-example</groupId>
       <artifactId>common</artifactId>
       <version>${project.version}</version>
     </dependency>
 
     <dependency>
-      <groupId>org.rawkintrevo.cylon</groupId>
+      <groupId>org.apache.mahout.cylon-example</groupId>
       <artifactId>frame-processors</artifactId>
       <version>${project.version}</version>
     </dependency>
@@ -129,7 +130,7 @@
           <archive>
             <manifest>
               <addClasspath>true</addClasspath>
-              <mainClass>org.rawkintrevo.cylon.CylonHTTPServer</mainClass>
+              <mainClass>org.apache.mahout.cylon-example.CylonHTTPServer</mainClass>
             </manifest>
           </archive>
           <descriptorRefs>
