+++ examples/cylons/examples/src/main/scala/org/rawkintrevo/cylon/examples/localfacialrecognition/SimpleFacialRecognitionApp.scala
@@ -1,6 +1,6 @@
-package org.rawkintrevo.cylon.examples.localfacialrecognition
+package org.apache.mahout.cylon-example.examples.localfacialrecognition
 
-import org.rawkintrevo.cylon.localengine.KafkaFaceDecomposer
+import org.apache.mahout.cylon-example.localengine.KafkaFaceDecomposer
 import org.slf4j.{Logger, LoggerFactory}
 
 /**
