+++ examples/cylons/flink-engine/pom.xml
@@ -20,9 +20,10 @@ under the License.
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>cylon</artifactId>
-    <groupId>org.rawkintrevo.cylon</groupId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>cylon-example</artifactId>
+    <groupId>org.apache.mahout</groupId>
+    <version>0.13.1-SNAPSHOT</version>
+    <relativePath>../</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
@@ -136,14 +137,14 @@ under the License.
 
     <!-- Our Stuff -->
     <dependency>
-      <groupId>org.rawkintrevo.cylon</groupId>
+      <groupId>org.apache.mahout.cylon-example</groupId>
       <artifactId>frame-processors</artifactId>
       <version>${project.version}</version>
     </dependency>
 
 
     <dependency>
-      <groupId>org.rawkintrevo.cylon</groupId>
+      <groupId>org.apache.mahout.cylon-example</groupId>
       <artifactId>common</artifactId>
       <version>${project.version}</version>
     </dependency>
@@ -334,7 +335,7 @@ under the License.
 
 							<transformers>
 								<transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-									<mainClass>org.rawkintrevo.cylon.flinkengine.apps.BasicFacialRecognition</mainClass>
+									<mainClass>org.apache.mahout.cylon-example.flinkengine.apps.BasicFacialRecognition</mainClass>
 								</transformer>
 							</transformers>
