+++ examples/cylons/flink-engine/src/main/scala/org/rawkintrevo/cylon/flinkengine/apps/BasicFacialRecognition.scala
@@ -1,4 +1,4 @@
-package org.rawkintrevo.cylon.flinkengine.apps
+package org.apache.mahout.cylon-example.flinkengine.apps
 
 /**
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -30,10 +30,10 @@ import org.apache.flink.streaming.api.windowing.time.Time
 import org.apache.flink.streaming.connectors.kafka.{FlinkKafkaConsumer010, FlinkKafkaProducer010}
 import org.apache.mahout.math.scalabindings._
 import org.apache.mahout.math.{Matrix, Vector}
-import org.rawkintrevo.cylon.flinkengine.coprocessfns.MarkupBufferedImagesWithDecomposedFacesCoProcessFunction
-import org.rawkintrevo.cylon.flinkengine.processfns.MergeImagesProcessFunction
-import org.rawkintrevo.cylon.flinkengine.schemas.{KeyedBufferedImageSchema, KeyedFrameBufferedImageSchema, MahoutVectorAndCoordsSchema}
-import org.rawkintrevo.cylon.flinkengine.windowfns._
+import org.apache.mahout.cylon-example.flinkengine.coprocessfns.MarkupBufferedImagesWithDecomposedFacesCoProcessFunction
+import org.apache.mahout.cylon-example.flinkengine.processfns.MergeImagesProcessFunction
+import org.apache.mahout.cylon-example.flinkengine.schemas.{KeyedBufferedImageSchema, KeyedFrameBufferedImageSchema, MahoutVectorAndCoordsSchema}
+import org.apache.mahout.cylon-example.flinkengine.windowfns._
 import org.slf4j.{Logger, LoggerFactory}
 
 // http://localhost:8090/cylon/cam/test-flink/flink-cluster-cam
