+++ examples/cylons/flink-engine/src/main/scala/org/rawkintrevo/cylon/flinkengine/apps/CountRectsByFrame.scala
@@ -1,4 +1,4 @@
-package org.rawkintrevo.cylon.flinkengine.apps
+package org.apache.mahout.cylon-example.flinkengine.apps
 
 /**
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -27,7 +27,7 @@ import org.apache.flink.streaming.api.windowing.time.Time
 import org.apache.flink.streaming.connectors.fs.bucketing.BucketingSink
 import org.apache.flink.streaming.connectors.kafka.{FlinkKafkaConsumer010, FlinkKafkaProducer010}
 import org.apache.mahout.math.Vector
-import org.rawkintrevo.cylon.flinkengine.schemas.{KeyedBufferedImageSchema, MahoutVectorAndCoordsSchema}
+import org.apache.mahout.cylon-example.flinkengine.schemas.{KeyedBufferedImageSchema, MahoutVectorAndCoordsSchema}
 import org.slf4j.{Logger, LoggerFactory}
 
 object CountRectsByFrame {
@@ -145,7 +145,7 @@ object CountRectsByFrame {
 //object FaceDetectorProcessor extends Serializable {
 //
 ////  //System.loadLibrary(Core.NATIVE_LIBRARY_NAME)
-////  Class.forName("org.rawkintrevo.cylon.opencv.LoadNative")
+////  Class.forName("org.apache.mahout.cylon-example.opencv.LoadNative")
 ////  //NativeUtils.loadOpenCVLibFromJar()
 ////
 ////  var inputRawImage: BufferedImage = _
