+++ examples/cylons/flink-engine/src/main/scala/org/rawkintrevo/cylon/flinkengine/processfns/MergeImagesProcessFunction.scala
@@ -1,4 +1,4 @@
-package org.rawkintrevo.cylon.flinkengine.processfns
+package org.apache.mahout.cylon-example.flinkengine.processfns
 
 import java.awt.image.BufferedImage
 import java.awt.Color
