+++ examples/cylons/flink-engine/src/main/scala/org/rawkintrevo/cylon/flinkengine/windowfns/FilterOutlierClusters.scala
@@ -1,4 +1,4 @@
-package org.rawkintrevo.cylon.flinkengine.windowfns
+package org.apache.mahout.cylon-example.flinkengine.windowfns
 
 import org.apache.flink.api.java.tuple.Tuple
 import org.apache.flink.streaming.api.scala.function.WindowFunction
