+++ examples/cylons/frameprocessors/pom.xml
@@ -1,13 +1,14 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <parent>
-    <artifactId>cylon</artifactId>
-    <groupId>org.rawkintrevo.cylon</groupId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>cylon-example</artifactId>
+    <groupId>org.apache.mahout</groupId>
+    <version>0.13.1-SNAPSHOT</version>
+    <relativePath>../</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <artifactId>frame-processors</artifactId>
-  <inceptionYear>2008</inceptionYear>
+  <inceptionYear>2017</inceptionYear>
   <properties>
   </properties>
 
@@ -55,7 +56,7 @@
 
 
     <dependency>
-      <groupId>org.rawkintrevo.cylon</groupId>
+      <groupId>org.apache.mahout.cylon-example</groupId>
       <artifactId>common</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
