+++ examples/cylons/frameprocessors/src/main/scala/org/rawkintrevo/cylon/frameprocessors/NativeUtils.scala
@@ -1,4 +1,4 @@
-package org.rawkintrevo.cylon.frameprocessors
+package org.apache.mahout.cylon-example.frameprocessors
 
 import java.io._
