+++ examples/cylons/httpserver/pom.xml
@@ -1,13 +1,14 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <parent>
-    <artifactId>cylon</artifactId>
-    <groupId>org.rawkintrevo.cylon</groupId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>cylon-example</artifactId>
+    <groupId>org.apache.mahout</groupId>
+    <version>0.13.1-SNAPSHOT</version>
+    <relativePath>../</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <artifactId>http-server</artifactId>
-  <inceptionYear>2008</inceptionYear>
+  <inceptionYear>2017</inceptionYear>
 
 
   <repositories>
@@ -111,7 +112,7 @@
           <archive>
             <manifest>
               <addClasspath>true</addClasspath>
-              <mainClass>org.rawkintrevo.cylon.CylonHTTPServer</mainClass>
+              <mainClass>org.apache.mahout.cylon-example.CylonHTTPServer</mainClass>
             </manifest>
           </archive>
           <descriptorRefs>
