+++ examples/cylons/local-engine/pom.xml
@@ -1,15 +1,34 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <!--
+   Licensed to the Apache Software Foundation (ASF) under one or more
+   contributor license agreements.  See the NOTICE file distributed with
+   this work for additional information regarding copyright ownership.
+   The ASF licenses this file to You under the Apache License, Version 2.0
+   (the "License"); you may not use this file except in compliance with
+   the License.  You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+  -->
+
+
   <parent>
-    <artifactId>cylon</artifactId>
-    <groupId>org.rawkintrevo.cylon</groupId>
-    <version>1.0-SNAPSHOT</version>
+    <artifactId>cylon-example</artifactId>
+    <groupId>org.apache.mahout</groupId>
+    <version>0.13.1-SNAPSHOT</version>
+    <relativePath>../</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <artifactId>localengine</artifactId>
   <name>${project.artifactId}</name>
   <description>My wonderfull scala app</description>
-  <inceptionYear>2015</inceptionYear>
+  <inceptionYear>2017</inceptionYear>
   <licenses>
     <license>
       <name>ASFv2</name>
@@ -54,13 +73,13 @@
     <dependency>
       <groupId>org.apache.mahout</groupId>
       <artifactId>mahout-math</artifactId>
-      <version>${mahout.version}</version>
+      <version>${project.version}</version>
     </dependency>
 
     <dependency>
       <groupId>org.apache.mahout</groupId>
       <artifactId>mahout-math-scala_${scala.compat.version}</artifactId>
-      <version>${mahout.version}</version>
+      <version>${project.version}</version>
     </dependency>
 
     <!-- Only for CLI Arg Parsing -->
@@ -71,19 +90,16 @@
     </dependency>
 
     <dependency>
-      <groupId>org.rawkintrevo.cylon</groupId>
+      <groupId>org.apache.mahout.cylon-example</groupId>
       <artifactId>frame-processors</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
 
     <dependency>
-      <groupId>org.rawkintrevo.cylon</groupId>
+      <groupId>org.apache.mahout.cylon-example</groupId>
       <artifactId>common</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
-
-
-
 
     <dependency>
       <groupId>org.apache.kafka</groupId>
@@ -130,30 +146,6 @@
           </includes>
         </configuration>
       </plugin>
-      <!-- Fat Jar -->
-      <!--<plugin>-->
-        <!--<artifactId>maven-assembly-plugin</artifactId>-->
-        <!--<configuration>-->
-          <!--<archive>-->
-            <!--<manifest>-->
-              <!--<addClasspath>true</addClasspath>-->
-              <!--<mainClass>org.rawkintrevo.cylon.App</mainClass>-->
-            <!--</manifest>-->
-          <!--</archive>-->
-          <!--<descriptorRefs>-->
-            <!--<descriptorRef>jar-with-dependencies</descriptorRef>-->
-          <!--</descriptorRefs>-->
-        <!--</configuration>-->
-        <!--<executions>-->
-          <!--<execution>-->
-            <!--<id>make-my-jar-with-dependencies</id>-->
-            <!--<phase>package</phase>-->
-            <!--<goals>-->
-              <!--<goal>single</goal>-->
-            <!--</goals>-->
-          <!--</execution>-->
-        <!--</executions>-->
-      <!--</plugin>-->
 
     </plugins>
   </build>
