+++ pom.xml
@@ -841,11 +841,11 @@
     <module>hdfs</module>
     <module>mr</module>
     <module>integration</module>
-    <module>examples</module>
     <module>distribution</module>
     <module>math-scala</module>
     <module>spark</module>
     <module>h2o</module>
+    <module>examples</module>
   </modules>
 
   <profiles>
